@@ -2,7 +2,7 @@ package com.example.emergensui.automotive_ui;
 
 import android.content.Intent;
 import android.os.Handler;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 
 public class splash_screen extends AppCompatActivity {
